+++ lib/rubocop/cop/platanus_cops.rb
@@ -2,5 +2,6 @@
 
 require_relative 'platanus/no_command'
 require_relative 'platanus/no_flash_in_api'
+require_relative 'platanus/no_enumerize'
 require_relative 'platanus/no_render_json'
 require_relative 'platanus/pundit_in_application_controller'