+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Features
 
 * Add cop Platanus/NoFlashInApi
+* Add cop Platanus/NoEnumerize
 * Add cop Platanus/NoRenderJson
 * Add cop Platanus/PunditInApplicationController